requirements.txt
discord.py
pyppeteer
pytz
